feat: replace needle with make-fetch-happen to leverage http caching
A common case will be re-fetching guidebook modules. The `make-fetch-happen` npm takes care of ETAG and other http cache protocols for us, as well as the disk caching, too.
