+++ package.json
@@ -82,6 +82,7 @@
     "@types/inquirer": "^8.2.1",
     "@types/js-yaml": "^4.0.5",
     "@types/json-diff": "^0.7.0",
+    "@types/make-fetch-happen": "^9.0.2",
     "@types/needle": "^2.5.3",
     "@types/node": "^17.0.31",
     "@types/unist": "^2.0.6",
@@ -106,6 +107,7 @@
   "dependencies": {
     "chalk": "^5.0.1",
     "debug": "^4.3.4",
+    "env-paths": "^3.0.0",
     "expand-home-dir": "^0.0.3",
     "figures": "^4.0.1",
     "front-matter": "^4.0.2",
@@ -114,8 +116,8 @@
     "hast-util-to-string": "^2.0.0",
     "inquirer": "^8.2.4",
     "js-yaml": "^4.1.0",
+    "make-fetch-happen": "^10.1.2",
     "mdast-util-to-markdown": "^1.3.0",
-    "needle": "^3.1.0",
     "ora": "^6.1.0",
     "remark-directive": "^2.0.1",
     "remark-frontmatter": "^4.0.1",
